No longer require param default pre-evaluation
By specifying the defaults as the values of variables used in `rcs::params`,
we need to inherit `rcs::params` inside `rcs`. This is useful in situations
where we have case statements or similar affecting the behaviour based on an
external fact - such as the running OS - but not that useful here.

Instead, and given that `rcs` contains notes about the default values, we
should just specify them here.

Fixes #2.