Explicitly upcase rcS option values
Values passed to options represented as strings inside /etc/default/rcS
should be capitalised as the default /etc/default/rcS file in Ubuntu has
the options capitalised. Options represented as integers shouldn't have
this method applied because it makes no sense to do so.

Fix this using the `upcase` method on the `string` object.