Forcibly downcase values
Originally, in #16, I thought that the default value was to upcase values in
this file, but since #18 was raised questioning this, I had another look on
a Vagrant VM built from the ubuntu/trusty64 box. The default in
/etc/default/rcS appears to be to use downcased values for the keys.

This commit enforces the use of downcased values regardless of how they're
passed in to this module.

Fixes #18.